conversion: section-create.js to tsx
